% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMetFunctions.R
\name{metrics}
\alias{metrics}
\title{A network metric calculation function}
\usage{
metrics(
  g,
  Edges,
  cutoff,
  clusterType = c("undirected", "global", "globalundirected", "localundirected", "local",
    "average", "localaverage", "localaverageundirected", "barrat", "weighted")
)
}
\arguments{
\item{g}{The network created by igraph}

\item{Edges}{A dataframe containing edge information between nodes}

\item{cutoff}{The cutoff value for calculating betweenness}

\item{clusterType}{The desired method for calcuating the clustering coefficient}
}
\description{
This function uses an igraph object and edges to calculate in-degree,
out-degreem in-strength, out-strength, eigenvector centrality, betweenness, and
clustering coefficients of all the nodes in the graph
}
\examples{
s1 <- dplyr::mutate(s1, Season = 1, sType = "Mating")
s2 <- dplyr::mutate(s2, Season = 2, sType = "Winter")
s3 <- dplyr::mutate(s3, Season = 3, sType = "Birth")
s1Edges <- data.frame(from = s1$"Actor", to = s1$"Recip", weight = s1$groom.rate)
s1G <- igraph::graph_from_data_frame(d = s1Edges, directed = TRUE)
s2Edges <- data.frame(from = s2$"Actor", to = s2$"Recip", weight = s2$groom.rate)
s2G <- igraph::graph_from_data_frame(d = s2Edges, directed = TRUE)
s3Edges <- data.frame(from = s3$"Actor", to = s3$"Recip", weight = s3$groom.rate)
s3G <- igraph::graph_from_data_frame(d = s3Edges, directed = TRUE)
met1 <- metrics(s1G, s1Edges, cutoff = -1, clusterType = "local")
met2 <- metrics(s2G, s2Edges, cutoff = -1, clusterType = "local")
met3 <- metrics(s3G, s3Edges, cutoff = -1, clusterType = "local")
}
\keyword{metrics}
