% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMet-package.R
\docType{package}
\name{netMet-package}
\alias{netMet}
\alias{netMet-package}
\title{netMet: Social Network Metric and Analysis}
\description{
The functions in this package calculates 7 unique social network metrics and allows the user to conduct various node based permutations of these metrics and visualize social networks as graphs which can view the relationship between multiple groups in a single graph. This package was built by Nico Jaworski, Blesing Apamaku, and Lemeng Li
}
\author{
\strong{Maintainer}: Nico Jaworski \email{nico.jaworski@utexas.edu}

}
\keyword{internal}
