% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netMetFunctions.R
\name{network}
\alias{network}
\title{Complex Social Network}
\usage{
network(
  edges,
  nodesInfo,
  label = c("Y", "N"),
  sexCol,
  speciesCol,
  male,
  redSpecies,
  diffSpecies = c("Y", "N")
)
}
\arguments{
\item{edges}{A data frame containing data dyads and the weight of the relationship}

\item{nodesInfo}{A data frame containing demographic data of each node, namely the node identity, sex, and species}

\item{label}{The user can choose if nodes will be labeled by their ID by setting label to "Y"}

\item{sexCol}{The name of the column in the nodesInfo data frame which contains sex designations}

\item{speciesCol}{The name of the column containing species names}

\item{male}{How the male sex is identified in the sex column, be it "M", "male", or otherwise}

\item{redSpecies}{The name of the species the user wants to be colored red}

\item{diffSpecies}{The user can set this to either "Y" or "N", with "Y" designating that there are different species in the data}
}
\description{
This function builds a social network from a series of edges provided by the user.
It is mean to graph the relationship between two or more distinct groups, such as
two species. Nodes are shaped according to the sex of the individual nodes, with males
being squares and females triangles. If there are 2 different groups or species, one is
colored red and the other green
}
\examples{
s1Edges <- data.frame(from = s1$"Actor", to = s1$"Recip", weight = s1$groom.rate)
node_info <- data.frame(node = elo$ID, sex = elo$Sex, species = c("A", "A", "A", "B", "B", "A", "A", "B", "B", "B", "A", "A", "A", "A", "A", "B", "A", "B", "B", "A", "A", "A", "A", "B", "A", "A"))
network(edges = s1Edges, nodesInfo = node_info, label = "Y", sexCol = "sex", speciesCol = "species", male = "M", redSpecies = "human", diffSpecies = "Y")
}
\keyword{network}
